Get logify* methods working again
The `@logger.logify*` decorators weren't getting wired up right.

Maybe due to updates in babel's decorator handling?  Not sure.

Changes necessary:

- Assign to `descriptor.value` instead of just returning function
- Reference the logger instance directly instead of as `this`

According to
[wycats](https://github.com/wycats/javascript-decorators#detailed-design)
a decorator is "an expression that evaluates to a function".  Guess it loses
its context binding in the process?

Anyway, this gets it working.

I also changed the log level to `silly`, and removed the (now obsolete)
extra `verbose` argument.