Eliminate intra-repo dep cycle and add maintainers docs
For now just requiring that `asini` and `asini-changelog` be installed
globally for use with the Asini repo.  Eventually the guts of the `asini`
package that `asini-changelog` uses should get pulled out into a dedicated
package (`asini-core`?) so we'll be able to eliminate that global install.
Might make sense to move the CLI into an `asini-cli` package, too, which would
eliminate _that_ global install.

For now, though, just documenting what it takes to publish Asini with globally
installed tools and removing the offending dep.